Sat Afternoon Changes
Resolved problem w/ determinisitc mice() MI due to not changing the seed for data creation in my simulation loop.

Added evaluation of parameter dataframe for all the parameters in MI, LD.

Added some randomness in the LD simulation by changing seed for data creation. This makes it so we don't get the same exact dataset everytime. But now my LD performs quite badly relative to MI...

As of now have two souces of randomness in imputations: (1) in data creation due to random seed (2) in missingness creation due to rbinom function being called seperately in each iteration. Try just putting either (1) or (2) in the "for" loop and keep the other constant? Maybe that'll affect LD/MI performance.